+++ pkg/scalers/rabbitmq_scaler_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
+	"time"
 )
 
 const (
@@ -45,6 +46,18 @@ var testRabbitMQMetadata = []parseRabbitMQMetadataTestData{
 	{map[string]string{"queueLength": "10", "queueName": "sample", "host": host, "protocol": "http"}, false, map[string]string{}},
 	// queue name with slashes
 	{map[string]string{"queueLength": "10", "queueName": "namespace/name", "hostFromEnv": host}, false, map[string]string{}},
+	// vhost passed
+	{map[string]string{"vhostName": "myVhost", "queueName": "namespace/name", "hostFromEnv": host}, false, map[string]string{}},
+	// vhost passed but empty
+	{map[string]string{"vhostName": "", "queueName": "namespace/name", "hostFromEnv": host}, false, map[string]string{}},
+	// protocol defined in authParams
+	{map[string]string{"queueName": "sample", "hostFromEnv": host}, false, map[string]string{"protocol": "http"}},
+	// auto protocol and a bad URL
+	{map[string]string{"queueName": "sample", "host": "something://"}, true, map[string]string{}},
+	// auto protocol and an HTTP URL
+	{map[string]string{"queueName": "sample", "host": "http://"}, false, map[string]string{}},
+	// auto protocol and an HTTPS URL
+	{map[string]string{"queueName": "sample", "host": "https://"}, false, map[string]string{}},
 }
 
 var rabbitMQMetricIdentifiers = []rabbitMQMetricIdentifier{
@@ -66,19 +79,20 @@ func TestRabbitMQParseMetadata(t *testing.T) {
 
 var testDefaultQueueLength = []parseRabbitMQMetadataTestData{
 	// use default queueLength
-	{map[string]string{"queueName": "sample", "host": host}, false, map[string]string{}},
+	{map[string]string{"queueName": "sample", "hostFromEnv": host}, false, map[string]string{}},
 	// use default queueLength with includeUnacked
-	{map[string]string{"queueName": "sample", "host": host, "protocol": "http"}, false, map[string]string{}},
+	{map[string]string{"queueName": "sample", "hostFromEnv": host, "protocol": "http"}, false, map[string]string{}},
 }
 
 func TestParseDefaultQueueLength(t *testing.T) {
 	for _, testData := range testDefaultQueueLength {
 		metadata, err := parseRabbitMQMetadata(&ScalerConfig{ResolvedEnv: sampleRabbitMqResolvedEnv, TriggerMetadata: testData.metadata, AuthParams: testData.authParams})
-		if err != nil && !testData.isError {
+		switch {
+		case err != nil && !testData.isError:
 			t.Error("Expected success but got error", err)
-		} else if testData.isError && err == nil {
+		case testData.isError && err == nil:
 			t.Error("Expected error but got success")
-		} else if metadata.queueLength != defaultRabbitMQQueueLength {
+		case metadata.queueLength != defaultRabbitMQQueueLength:
 			t.Error("Expected default queueLength =", defaultRabbitMQQueueLength, "but got", metadata.queueLength)
 		}
 	}
@@ -88,84 +102,116 @@ type getQueueInfoTestData struct {
 	response       string
 	responseStatus int
 	isActive       bool
+	extraMetadata  map[string]string
+	vhostPath      string
 }
 
 var testQueueInfoTestData = []getQueueInfoTestData{
-	{`{"messages": 4, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true},
-	{`{"messages": 1, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true},
-	{`{"messages": 1, "messages_unacknowledged": 0, "name": "evaluate_trials"}`, http.StatusOK, true},
-	{`{"messages": 0, "messages_unacknowledged": 0, "name": "evaluate_trials"}`, http.StatusOK, false},
-	{`Password is incorrect`, http.StatusUnauthorized, false},
+	{`{"messages": 4, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true, nil, ""},
+	{`{"messages": 1, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true, nil, ""},
+	{`{"messages": 1, "messages_unacknowledged": 0, "name": "evaluate_trials"}`, http.StatusOK, true, nil, ""},
+	{`{"messages": 0, "messages_unacknowledged": 0, "name": "evaluate_trials"}`, http.StatusOK, false, nil, ""},
+	{`Password is incorrect`, http.StatusUnauthorized, false, nil, ""},
 }
 
 var vhostPathes = []string{"/myhost", "", "/", "//", "/%2F"}
 
+var testQueueInfoTestDataSingleVhost = []getQueueInfoTestData{
+	{`{"messages": 4, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true, map[string]string{"hostFromEnv": "plainHost", "vhostName": "myhost"}, "/myhost"},
+	{`{"messages": 4, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true, map[string]string{"hostFromEnv": "plainHost", "vhostName": "/"}, "/"},
+	{`{"messages": 4, "messages_unacknowledged": 1, "name": "evaluate_trials"}`, http.StatusOK, true, map[string]string{"hostFromEnv": "plainHost", "vhostName": ""}, "/"},
+}
+
 func TestGetQueueInfo(t *testing.T) {
+	allTestData := []getQueueInfoTestData{}
 	for _, testData := range testQueueInfoTestData {
-		testData := testData
 		for _, vhostPath := range vhostPathes {
-			expectedVhost := "myhost"
+			testData := testData
+			testData.vhostPath = vhostPath
+			allTestData = append(allTestData, testData)
+		}
+	}
+	allTestData = append(allTestData, testQueueInfoTestDataSingleVhost...)
 
-			if vhostPath != "/myhost" {
-				expectedVhost = "%2F"
+	for _, testData := range allTestData {
+		testData := testData
+		expectedVhost := "myhost"
+
+		if testData.vhostPath != "/myhost" {
+			expectedVhost = "%2F"
+		}
+
+		var apiStub = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			expectedPath := "/api/queues/" + expectedVhost + "/evaluate_trials"
+			if r.RequestURI != expectedPath {
+				t.Error("Expect request path to =", expectedPath, "but it is", r.RequestURI)
 			}
 
-			var apiStub = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				expectedPath := "/api/queues/" + expectedVhost + "/evaluate_trials"
-				if r.RequestURI != expectedPath {
-					t.Error("Expect request path to =", expectedPath, "but it is", r.RequestURI)
-				}
-
-				w.WriteHeader(testData.responseStatus)
-				w.Write([]byte(testData.response))
-			}))
-
-			resolvedEnv := map[string]string{host: fmt.Sprintf("%s%s", apiStub.URL, vhostPath)}
-
-			metadata := map[string]string{
-				"queueLength": "10",
-				"queueName":   "evaluate_trials",
-				"hostFromEnv": host,
-				"protocol":    "http",
+			w.WriteHeader(testData.responseStatus)
+			_, err := w.Write([]byte(testData.response))
+			if err != nil {
+				t.Error("Expect request path to =", testData.response, "but it is", err)
 			}
+		}))
 
-			s, err := NewRabbitMQScaler(&ScalerConfig{ResolvedEnv: resolvedEnv, TriggerMetadata: metadata, AuthParams: map[string]string{}})
+		resolvedEnv := map[string]string{host: fmt.Sprintf("%s%s", apiStub.URL, testData.vhostPath), "plainHost": apiStub.URL}
 
+		metadata := map[string]string{
+			"queueLength": "10",
+			"queueName":   "evaluate_trials",
+			"hostFromEnv": host,
+			"protocol":    "http",
+		}
+		for k, v := range testData.extraMetadata {
+			metadata[k] = v
+		}
+
+		s, err := NewRabbitMQScaler(
+			&ScalerConfig{
+				ResolvedEnv:       resolvedEnv,
+				TriggerMetadata:   metadata,
+				AuthParams:        map[string]string{},
+				GlobalHTTPTimeout: 1000 * time.Millisecond,
+			},
+		)
+
+		if err != nil {
+			t.Error("Expect success", err)
+		}
+
+		ctx := context.TODO()
+		active, err := s.IsActive(ctx)
+
+		if testData.responseStatus == http.StatusOK {
 			if err != nil {
 				t.Error("Expect success", err)
 			}
 
-			ctx := context.TODO()
-			active, err := s.IsActive(ctx)
-
-			if testData.responseStatus == http.StatusOK {
-				if err != nil {
-					t.Error("Expect success", err)
-				}
-
-				if active != testData.isActive {
-					if testData.isActive {
-						t.Error("Expect to be active")
-					} else {
-						t.Error("Expect to not be active")
-					}
-				}
-			} else {
-				if !strings.Contains(err.Error(), testData.response) {
-					t.Error("Expect error to be like '", testData.response, "' but it's '", err, "'")
+			if active != testData.isActive {
+				if testData.isActive {
+					t.Error("Expect to be active")
+				} else {
+					t.Error("Expect to not be active")
 				}
 			}
+		} else if !strings.Contains(err.Error(), testData.response) {
+			t.Error("Expect error to be like '", testData.response, "' but it's '", err, "'")
 		}
 	}
 }
 
 func TestRabbitMQGetMetricSpecForScaling(t *testing.T) {
 	for _, testData := range rabbitMQMetricIdentifiers {
-		meta, err := parseRabbitMQMetadata(&ScalerConfig{ResolvedEnv: map[string]string{"myHostSecret": "myHostSecret"}, TriggerMetadata: testData.metadataTestData.metadata, AuthParams: nil})
+		meta, err := parseRabbitMQMetadata(&ScalerConfig{ResolvedEnv: sampleRabbitMqResolvedEnv, TriggerMetadata: testData.metadataTestData.metadata, AuthParams: nil})
 		if err != nil {
 			t.Fatal("Could not parse metadata:", err)
 		}
-		mockRabbitMQScaler := rabbitMQScaler{meta, nil, nil}
+		mockRabbitMQScaler := rabbitMQScaler{
+			metadata:   meta,
+			connection: nil,
+			channel:    nil,
+			httpClient: http.DefaultClient,
+		}
 
 		metricSpec := mockRabbitMQScaler.GetMetricSpecForScaling()
 		metricName := metricSpec[0].External.Metric.Name